Fixed various PMU compiler warnings
[Defect]
MobC00378016

[Problem]
Various compiler warnings are observed in PMU-related files

[Solution]
Fix each compiler warning

[Reviewers]

[Testing]

[Test Cases]
- Compile tested - warnings resolved in reported files
- Phone boot up successfully

[Label/CL# Synced To]
sdb-linux-3.10.0-hawaii-java

[Platform]
Hawaii Garnet EDN020

[BROADCOM INTERNAL NOTES]

Reviewed-on: http://mps-gerrit.broadcom.com/43131
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>

+++ drivers/rtc/rtc-bcmpmu59xxx.c
@@ -128,7 +128,7 @@ static int bcmpmu_alarm_irq_enable(struct device *dev,
 		 * firing at unwanted time.
 		 */
 		ret = rdata->bcmpmu->write_dev(rdata->bcmpmu, PMU_REG_RTCYR_A1,
-					0xFF);
+					(u8)0xFF);
 		if (unlikely(ret))
 			goto err;
